Initial rough code check-in
This is my initial, written-at-1:30AM-because-I-couldn't-sleep-despite-having-to-be-up-in-5-hours code. It is not pretty. But it works.

Unfortunately I do not know the source for wordsEn.txt so I don't know it's licence, therefore I have not included it. It's a simple wordlist, many like it exist so finding one should be easy